fixed some stuff on colormap

+++ viewer/MainWindow.cpp
@@ -439,10 +439,10 @@ void MainWindow::lowerThresholdChanged( int lowerThreshold )
 {
 
 	if( m_ViewerCore->getCurrentImage()->getImageProperties().imageType == ImageHolder::z_map ) {
-		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<float>( "lowerThreshold", ( m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>() / 1000 ) * ( lowerThreshold  )  );
+		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<double>( "lowerThreshold", ( m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>() / 1000 ) * ( lowerThreshold  )  );
 	} else {
 		double range = m_ViewerCore->getCurrentImage()->getMinMax().second->as<double>() - m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>();
-		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<float>( "lowerThreshold", ( range / 1000 ) * ( lowerThreshold  ) + m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>() );
+		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<double>( "lowerThreshold", ( range / 1000 ) * ( lowerThreshold  ) + m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>() );
 	}
 	m_ViewerCore->updateScene();
 }
@@ -450,12 +450,11 @@ void MainWindow::lowerThresholdChanged( int lowerThreshold )
 void MainWindow::upperThresholdChanged( int upperThreshold )
 {
 	if( m_ViewerCore->getCurrentImage()->getImageProperties().imageType == ImageHolder::z_map ) {
-		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<float>( "upperThreshold", (  m_ViewerCore->getCurrentImage()->getMinMax().second->as<double>() / 1000 ) * ( upperThreshold  ) );
+		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<double>( "upperThreshold", (  m_ViewerCore->getCurrentImage()->getMinMax().second->as<double>() / 1000 ) * ( upperThreshold  ) );
 	} else {
 		double range = m_ViewerCore->getCurrentImage()->getMinMax().second->as<double>() - m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>();
-		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<float>( "upperThreshold", ( range / 1000 ) * ( upperThreshold + 1 )  + m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>() ) ;
+		m_ViewerCore->getCurrentImage()->getPropMap().setPropertyAs<double>( "upperThreshold", ( range / 1000 ) * ( upperThreshold + 1 )  + m_ViewerCore->getCurrentImage()->getMinMax().first->as<double>() ) ;
 	}
-
 	m_ViewerCore->updateScene();
 
 }
